Make double

+++ lib/screens/book_an_appointment/patient_booking_screen.dart
@@ -18,7 +18,7 @@ class PatientBookingScreen extends StatefulWidget {
   final String? dateText;
   final String? timeSlotText;
   final String id;
-  final int? consultationFee;
+  final double? consultationFee;
   final String? title;
   final String? doctorFullName;
   final String? locationName;
